+++ docs/source/conf.py
@@ -25,7 +25,7 @@ import django
 
 django.setup()
 
-__version__ = '0.5'
+__version__ = '0.7'
 
 # -- Project information -----------------------------------------------------