+++ tests/test_portfolio/tests_models.py
@@ -21,15 +21,17 @@
 
 from django.test import TestCase
 
+from portfolio.Counterparty import Counterparty
 from portfolio.models import AreaSource, MultiAreaSource, PointSource
 
 from portfolio.EmissionsSource import BuildingEmissionsSource, EmissionsSource, GPCEmissionsSource, GPPEmissionsSource
 
 from portfolio.Borrower import Borrower
-from portfolio.Asset import Building, ProjectAsset
+from portfolio.Asset import Building, PowerPlant
 from portfolio.Contractor import Contractor
 from portfolio.Loan import Loan
 from portfolio.Swap import Swap
+from portfolio.Certificate import Certificate
 from portfolio.Mortgage import Mortgage
 from portfolio.Operator import Operator
 from portfolio.PortfolioManager import PortfolioManager
@@ -102,6 +104,11 @@ class PortfolioModelTests(TestCase):
         instance = Contractor.objects.get()
         self.assertEquals("1", str(instance))
 
+    def test_counterparty_str(self):
+        Counterparty.objects.create(counterparty_identifier=1)
+        instance = Counterparty.objects.get()
+        self.assertEquals("1", str(instance))
+
     def test_limit_structure_str(self):
         LimitStructure.objects.create(name='test')
         instance = LimitStructure.objects.get()
@@ -117,9 +124,15 @@ class PortfolioModelTests(TestCase):
         instance = Swap.objects.get()
         self.assertEquals("test", str(instance))
 
-    def test_swap_str(self):
-        Swap.objects.create(swap_identifier='test')
-        instance = Swap.objects.get()
+
+    def test_certificate_str(self):
+        Certificate.objects.create(certificate_identifier='test')
+        instance = Certificate.objects.get()
+        self.assertEquals("test", str(instance))
+
+    def test_power_plant_str(self):
+        PowerPlant.objects.create(production_device_number='test')
+        instance = PowerPlant.objects.get()
         self.assertEquals("test", str(instance))
 
     def test_mortgage_str(self):