+++ portfolio/CHANGELOG.rst
@@ -1,6 +1,11 @@
 ChangeLog
 ===========================
 
+v0.4.0 (03-06-2023)
+-------------------
+* Model testing framework
+* Certificate data model for EECS certificates and Guarantees of Origin
+
 v0.3.0 (21-04-2022)
 -------------------
 * Earth Day 2022 release (starting separate changelog for this app)