+++ reporting/CHANGELOG.rst
@@ -1,6 +1,10 @@
 ChangeLog
 ===========================
 
+v0.6.0 (03-06-2023)
+-------------------
+* Model testing framework
+
 v0.5.0 (21-06-2022)
 -------------------
 * Aggregated Statistics Model